Changed make to use linked.ll

+++ experimentation/05_pass/tinyAES.c
@@ -151,7 +151,7 @@ static uint8_t getSBoxInvert(uint8_t num) { return rsbox[num]; }
 // This function produces Nb(Nr+1) round keys. The round keys are used in each
 // round to decrypt the states.
 static void KeyExpansion(void) {
-  uint32_t i, j, k;
+  uint8_t i, j, k;
   uint8_t tempa[4]; // Used for the column/row operations
 
   // The first round key is the key itself.
